feat: Implement paginated fetching of GitLab merge request diffs
- Added loop-based pagination to retrieve all MR diffs using `x-next-page` header.
- Introduced `GitlabDiffResponseModel` to map diff response structure.
- Improved logging and error handling for better traceability.
- Updated tests to reflect new diff fetching logic.
